refactor file and class structure in schemas
