+++ athena/athena/schemas/feedback/graded_feedback/__init__.py
@@ -1,0 +1,4 @@
+from .graded_feedback import GradedFeedback
+from .graded_programming_feedback import GradedProgrammingFeedback
+from .graded_modelling_feedback import GradedModellingFeedback
+from .graded_text_feedback import GradedTextFeedback
