+++ athena/athena/schemas/feedback/non_graded_feedback/__init__.py
@@ -1,0 +1,2 @@
+from .non_graded_feedback import NonGradedFeedback
+from .non_graded_programming_feedback import NonGradedProgrammingFeedback
